avaliacao_de_desempenho/requirements.txt
psycopg2-binary==2.9.9
pandas==2.1.4
tabulate==0.9.0 